simplify the code further
