+++ pkg/sink/mysql/mysql_writer.go
@@ -26,6 +26,7 @@ import (
 	commonEvent "github.com/pingcap/ticdc/pkg/common/event"
 	"github.com/pingcap/ticdc/pkg/errors"
 	"github.com/pingcap/ticdc/pkg/metrics"
+	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 )
 
@@ -305,4 +306,32 @@ func (w *Writer) Close() {
 		w.cancel()
 	}
 	w.dmlSession.close(w)
+}
+
+type rowsAffectedLabels struct {
+	countType string
+	rowType   string
+}
+
+func (w *Writer) recordTotalRowsAffected(actualRowsAffected, expectedRowsAffected int64) {
+	w.getRowsAffectedCounter("actual", "total").Add(float64(actualRowsAffected))
+	w.getRowsAffectedCounter("expected", "total").Add(float64(expectedRowsAffected))
+}
+
+func (w *Writer) recordRowsAffected(rowsAffected int64, rowType common.RowType) {
+	w.getRowsAffectedCounter("actual", rowType.String()).Add(float64(rowsAffected))
+	w.getRowsAffectedCounter("expected", rowType.String()).Add(1)
+	w.recordTotalRowsAffected(rowsAffected, 1)
+}
+
+func (w *Writer) getRowsAffectedCounter(countType, rowType string) prometheus.Counter {
+	labels := rowsAffectedLabels{countType: countType, rowType: rowType}
+	counter, loaded := w.rowsAffectedCounters.Load(labels)
+	if !loaded {
+		counter := execDMLEventRowsAffectedCounter.WithLabelValues(
+			w.ChangefeedID.Keyspace(), w.ChangefeedID.Name(), countType, rowType)
+		w.rowsAffectedCounters.Store(labels, counter)
+		return counter
+	}
+	return counter.(prometheus.Counter)
 }
